requirements.txt
ultralyticsplus~=0.1.0
torch~=2.2.2
torchvision~=0.17.2
matplotlib~=3.8.4
numpy~=1.26.4
pillow~=10.3.0